Actualización del e-commerce en el backend

+++ controllers/productoController.js
@@ -1,51 +1,47 @@
-import Producto from '../models/Producto.js';
-import fs from 'fs-extra';
-
+import Producto from "../models/Producto.js";
+import fs from "fs-extra";
 import {
+
    uploadImage,
    deleteImage
-} from '../helper/cloudinary.js';
 
+} from '../helper/cloudinary.js';
 const prueba = (req, res) => {
    res.send({
       msg: "En esta ruta gestionaremos todas las peticiones correspondiente al modelo de Producto"
    })
 };
-
 const createProductos = async (req, res) => {
    try {
-
       const { nombre, description, precio, stock } = req.body;
       let image;
-
-      if (req.files.image) {
-
-         const result = await uploadImage(req.files.image.tempFilePath);
-         await fs.remove(req.files.image.tempFilePath);
-
+      if (req.files !== null) {
+         if (req.files.image) {
+            const result = await uploadImage(req.files.image.tempFilePath);
+            await fs.remove(req.files.image.tempFilePath);
+            image = {
+               url: result.secure_url,
+               public_id: result.public_id,
+            };
+            //console.log(result);
+         }
+      } else {
          image = {
-            url: result.secure_url,
-            public_id: result.public_id,
+            url: "https://res.cloudinary.com/dgi1aoxog/image/upload/v1668981098/productos/productoblacoynegro_ty6jw2.jpg",
+            public_id: "productos/productoblacoynegro_ty6jw2",
          };
-
-         console.log(result);
       }
-
       const Newproducto = new Producto({ nombre, description, precio, image, stock });
       await Newproducto.save();
-
       return res.json(Newproducto);
    } catch (error) {
       console.log(error);
       return res.status(500).json({ msg: error.message });
    }
 };
-
 const getProducto = async (req, res) => {
    try {
-
       const OneProduct = await Producto.findById(req.params.id);
-
       if (!OneProduct) {
          return res.sendStatus(404);
       } else {
@@ -55,7 +51,6 @@ const getProducto = async (req, res) => {
       return res.status(500).json({ message: error.message });
    }
 };
-
 const getProductos = async (req, res) => {
    try {
       const productos = await Producto.find();
@@ -65,35 +60,45 @@ const getProductos = async (req, res) => {
       return res.status(500).json({ message: error.message });
    }
 };
-
 const updateProductos = async (req, res) => {
-
    const { id, nombre, description, precio, stock } = req.body;
    try {
       const updateProducto = await Producto.findById(id);
-      // console.log(updatePlato);
+      //console.log(updateProducto);
+
       updateProducto.nombre = nombre;
       updateProducto.description = description;
       updateProducto.precio = precio;
-      updateProducto.precio = stock;
-      
-      if(req.files!==null){
+      updateProducto.stock = stock;
+
+      if (req.files !== null) {
          if (req.files.image) {
-            await deleteImage(updateProducto.image.public_id);
+            if (updateProducto.image.public_id && updateProducto.image.public_id !== "productos/productoblacoynegro_ty6jw2") {
+               await deleteImage(updateProducto.image.public_id);
+            }
             const result = await uploadImage(req.files.image.tempFilePath);
             await fs.remove(req.files.image.tempFilePath);
-   
             updateProducto.image = {
                url: result.secure_url,
                public_id: result.public_id,
             };
-         
+
+         } else {
+            updateProducto.image = {
+               url: "https://res.cloudinary.com/dgi1aoxog/image/upload/v1668981098/productos/productoblacoynegro_ty6jw2.jpg",
+               public_id: "productos/productoblacoynegro_ty6jw2",
+            };
+         }
+      } else {
+         if (!updateProducto.image.public_id) {
+            updateProducto.image = {
+               url: "https://res.cloudinary.com/dgi1aoxog/image/upload/v1668981098/productos/productoblacoynegro_ty6jw2.jpg",
+               public_id: "productos/productoblacoynegro_ty6jw2",
+            };
+         }
       }
-      
-         await updateProducto.save();
-         
-         return res.status(204).json(updatePlato);
-      }
+      await updateProducto.save();
+      return res.status(204).json(updateProducto);
    } catch (error) {
       console.log(error.message);
    }
@@ -102,17 +107,17 @@ const updateProductos = async (req, res) => {
 const deleteProductos = async (req, res) => {
    try {
       const productRemoved = await Producto.findByIdAndDelete(req.params.id);
-
       if (!productRemoved) {
-         const error = new Error("Token no valido");
+         //const error = new Error("Token no valido");
          return res.sendStatus(404);
       } else {
          if (productRemoved.image.public_id) {
             await deleteImage(productRemoved.image.public_id);
          }
-         return res.sendStatus(204);
+         return res.status(200).json({
+            msg: "Producto Eliminado exitosamente"
+         });
       }
-
    } catch (error) {
       return res.status(500).json({ message: error.message });
    }
@@ -120,8 +125,9 @@ const deleteProductos = async (req, res) => {
 
 export {
    prueba,
-   createProductos,
    getProductos,
+   createProductos,
    updateProductos,
-   deleteProductos
+   deleteProductos,
+   getProducto
 };
